git-svn-id: http://vps6982.inmotionhosting.com/svn/lasi@790 49725315-0998-4498-bb4a-373bc1a52cb2

+++ LASI_FileSystem/TaggerAndFileAbstraction/FileManager.cs
@@ -1,12 +1,12 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Diagnostics;
 using System.IO;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using LASI.FileSystem.FileTypes;
 using SharpNLPTaggingModule;
+using LASI.Utilities;
 namespace LASI.FileSystem
 {
     /// <summary>
@@ -92,33 +92,31 @@ namespace LASI.FileSystem
         /// </summary>
         /// <param name="filePath">A partial or full, extensionless or extensionful, file newPath containing the name of the file to check.</param>
         /// <returns>False if a file with the same name, irrespective of its extension, is part of the project. False otherwise.</returns>
-        public static bool FileInProject(string filePath) {
+        public static bool HasSimilarFile(string filePath) {
             var fileName = new string(
                 filePath.Reverse().
                 SkipWhile(c => c != '.').
                 Skip(1).TakeWhile(c => c != '\\').
                 Reverse().ToArray());
-            return !FileInProjectSet(fileName);
+            return !localDocumentNames.Contains(fileName);
         }
         /// <summary>
         /// Returns a value indicating whether a file with the same name as that of the given InputFile, irrespective of its extension, is part of the project. 
         /// </summary>
         /// <param name="filePath">An an Instance of the InputFile class or one of its descendents.</param>
         /// <returns>False if a file with the same name, irrespective of it'subject extension, is part of the project. False otherwise.</returns>
-        public static bool FileInProject(InputFile inputFile) {
-            return !FileInProjectSet(inputFile.NameSansExt);
+        public static bool HasSimilarFile(InputFile inputFile) {
+            return !localDocumentNames.Contains(inputFile.NameSansExt);
         }
-
-        private static bool FileInProjectSet(string fileName) {
-            return localDocumentNames.Contains(fileName);
-        }
-
 
 
         /// <summary>
         /// Performs the necessary conversions, based on the format of all files within the project.
         /// </summary>
         public static void ConvertAsNeeded() {
+            if (!Initialized) {
+                throw new FileManagerNotInitializedException();
+            }
             ConvertDocFiles();
             ConvertDocxToText();
             TagTextFiles();
@@ -128,6 +126,9 @@ namespace LASI.FileSystem
         /// Asynchronously performs the necessary conversions, based on the format of all files within the project.
         /// </summary>
         public static async Task ConvertAsNeededAsync() {
+            if (!Initialized) {
+                throw new FileManagerNotInitializedException();
+            }
             await Task.WhenAll(
                 ConvertDocxToTextAsync(),
                 ConvertPdfFilesAsync()
@@ -165,6 +166,9 @@ namespace LASI.FileSystem
         /// </summary>
         /// <param name="filesToKeep">collction of file path strings indicating which files are not to be culled. All others will summarilly executed.</param>
         public static void RemoveAllNotIn(IEnumerable<string> filesToKeep) {
+            if (!Initialized) {
+                throw new FileManagerNotInitializedException();
+            }
             RemoveAllNotIn(from f in filesToKeep
                            select f.IndexOf('.') > 0 ? WrapperMap[f.Substring(f.LastIndexOf('.'))](f) : new TextFile(f));
         }
@@ -173,6 +177,9 @@ namespace LASI.FileSystem
         /// </summary>
         /// <param name="filesToKeep">collection of InputFile objects indicating which files are not to be culled. All others will summarilly executed.</param>
         public static void RemoveAllNotIn(IEnumerable<InputFile> filesToKeep) {
+            if (!Initialized) {
+                throw new FileManagerNotInitializedException();
+            }
             var toRemove = from f in localDocumentNames
                            where (from k in filesToKeep
                                   where f == k.NameSansExt
@@ -188,9 +195,9 @@ namespace LASI.FileSystem
             docFiles.RemoveAll(f => f.NameSansExt == fileName);
             docXFiles.RemoveAll(f => f.NameSansExt == fileName);
             pdfFiles.RemoveAll(f => f.NameSansExt == fileName);
-            taggedFiles = new System.Collections.Concurrent.ConcurrentBag<TaggedFile>(from f in taggedFiles
-                                                                                      where f.NameSansExt != fileName
-                                                                                      select f);
+            taggedFiles = (from f in taggedFiles
+                           where f.NameSansExt != fileName
+                           select f).ToList();
         }
         /// <summary>
         /// Adds the document indicated by the specified path string to the project
@@ -199,6 +206,9 @@ namespace LASI.FileSystem
         /// <param name="overwrite">True to overwrite existing documents within the project with the same name, False otherwise. Defaults to False</param>
         /// <returns>An InputFile object which acts as a wrapper around the project relative path of the newly added file.</returns>
         public static InputFile AddFile(string path, bool overwrite = true) {
+            if (!Initialized) {
+                throw new FileManagerNotInitializedException();
+            }
             var ext = path.Substring(path.LastIndexOf('.')).ToLower();
             try {
                 var originalFile = FileManager.WrapperMap[ext](path);
@@ -227,11 +237,17 @@ namespace LASI.FileSystem
         /// </summary>
         /// <param name="file">The document to remove</param>
         public static void RemoveFile(InputFile file) {
+            if (!Initialized) {
+                throw new FileManagerNotInitializedException();
+            }
             RemoveAllAlikeFiles(file.NameSansExt);
 
             RemoveFile(file as dynamic);
         }
         public static void RemoveFile(string file) {
+            if (!Initialized) {
+                throw new FileManagerNotInitializedException();
+            }
             RemoveAllAlikeFiles(file);
         }
 
@@ -254,6 +270,9 @@ namespace LASI.FileSystem
         /// </summary>
         /// <param name="files">0 or more instances of the DocFile class which encapsulate .doc files</param>
         public static void ConvertDocFiles(params DocFile[] files) {
+            if (!Initialized) {
+                throw new FileManagerNotInitializedException();
+            }
             if (files.Length == 0)
                 files = docFiles.ToArray();
             foreach (var doc in from d in files
@@ -274,6 +293,9 @@ namespace LASI.FileSystem
         /// </summary>
         /// <param name="files">0 or more instances of the DocFile class which encapsulate .doc files</param>
         public static async Task ConvertDocFilesAsync(params DocFile[] files) {
+            if (!Initialized) {
+                throw new FileManagerNotInitializedException();
+            }
             if (files.Length == 0)
                 files = docFiles.ToArray();
             foreach (var doc in from d in files
@@ -288,6 +310,9 @@ namespace LASI.FileSystem
         }
 
         public static void ConvertPdfToText(params PdfFile[] files) {
+            if (!Initialized) {
+                throw new FileManagerNotInitializedException();
+            }
             if (files.Length == 0)
                 files = pdfFiles.ToArray();
             foreach (var pdf in from file in files
@@ -303,6 +328,9 @@ namespace LASI.FileSystem
         }
 
         public static async Task ConvertPdfFilesAsync(params PdfFile[] files) {
+            if (!Initialized) {
+                throw new FileManagerNotInitializedException();
+            }
             if (files.Length == 0)
                 files = pdfFiles.ToArray();
             foreach (var pdf in from file in files
@@ -325,6 +353,9 @@ namespace LASI.FileSystem
         /// </summary>
         /// <param name="files">0 or more instances of the DocXFile class which encapsulate .docx files</param>
         public static void ConvertDocxToText(params DocXFile[] files) {
+            if (!Initialized) {
+                throw new FileManagerNotInitializedException();
+            }
             if (files.Length == 0)
                 files = docXFiles.ToArray();
             foreach (var doc in from d in files
@@ -345,6 +376,9 @@ namespace LASI.FileSystem
         /// </summary>
         /// <param name="files">0 or more instances of the DocXFile class which encapsulate .docx files</param>
         public static async Task ConvertDocxToTextAsync(params DocXFile[] files) {
+            if (!Initialized) {
+                throw new FileManagerNotInitializedException();
+            }
             if (files.Length == 0)
                 files = docXFiles.ToArray();
             foreach (var doc in from d in files
@@ -370,6 +404,9 @@ namespace LASI.FileSystem
         /// </summary>
         /// <param name="files">0 or more instances of the TextFile class which encapsulate text files</param>
         public static void TagTextFiles(params TextFile[] files) {
+            if (!Initialized) {
+                throw new FileManagerNotInitializedException();
+            }
             if (files.Length == 0)
                 files = textFiles.ToArray();
             foreach (var doc in from d in files
@@ -391,6 +428,9 @@ namespace LASI.FileSystem
         /// </summary>
         /// <param name="files">0 or more instances of the TextFile class which encapsulate text files</param>
         public static async Task TagTextFilesAsync(params TextFile[] files) {
+            if (!Initialized) {
+                throw new FileManagerNotInitializedException();
+            }
             if (files.Length == 0)
                 files = textFiles.ToArray();
             var tasks = (from d in files
@@ -416,6 +456,9 @@ namespace LASI.FileSystem
         /// Copies the entire contents of the current project directory to a predetermined, relative path
         /// </summary>
         public static void BackupProject() {
+            if (!Initialized) {
+                throw new FileManagerNotInitializedException();
+            }
             var projd = new DirectoryInfo(ProjectDir);
             var pard = new DirectoryInfo(projd.Parent.FullName);
             var desitination = Directory.CreateDirectory(pard.FullName + "\\backup\\" + ProjectName);
@@ -426,7 +469,16 @@ namespace LASI.FileSystem
             }
         }
         public static void DecimateProject() {
-            Directory.Delete(ProjectDir, true);
+            if (!Initialized) {
+                throw new FileManagerNotInitializedException();
+            }
+            try {
+                Directory.Delete(ProjectDir, true);
+            }
+            catch (DirectoryNotFoundException e) {
+                Output.WriteLine(e.Message);
+                Output.WriteLine("Directory could not be found for forced cleabup");
+            }
         }
         #endregion
 
@@ -515,7 +567,7 @@ namespace LASI.FileSystem
         /// </summary>
         public static IReadOnlyList<TextFile> TextFiles {
             get {
-                return FileManager.textFiles;
+                return textFiles;
             }
         }
         /// <summary>
@@ -524,7 +576,7 @@ namespace LASI.FileSystem
         /// </summary>
         public static IReadOnlyList<DocXFile> DocXFiles {
             get {
-                return FileManager.docXFiles;
+                return docXFiles;
             }
         }
         /// <summary>
@@ -533,7 +585,7 @@ namespace LASI.FileSystem
         /// </summary>
         public static IReadOnlyList<DocFile> DocFiles {
             get {
-                return FileManager.DocFiles;
+                return DocFiles;
             }
         }
         /// <summary>
@@ -542,10 +594,15 @@ namespace LASI.FileSystem
         /// </summary>
         public static IReadOnlyList<PdfFile> PdfFiles {
             get {
-                return FileManager.pdfFiles;
+                return pdfFiles;
             }
         }
 
+        public static IReadOnlyList<TaggedFile> TaggedFiles {
+            get {
+                return taggedFiles;
+            }
+        }
 
         public static readonly WrapperDict WrapperMap = new WrapperDict();
 
@@ -556,57 +613,24 @@ namespace LASI.FileSystem
 
         public static bool Initialized {
             get;
-            set;
+            private set;
         }
 
-        static HashSet<string> localDocumentNames = new HashSet<string>();
+        private static HashSet<string> localDocumentNames = new HashSet<string>();
 
-        static List<DocFile> docFiles = new List<DocFile>();
+        private static List<DocFile> docFiles = new List<DocFile>();
 
-        static List<DocXFile> docXFiles = new List<DocXFile>();
+        private static List<DocXFile> docXFiles = new List<DocXFile>();
 
-        static List<PdfFile> pdfFiles = new List<PdfFile>();
+        private static List<PdfFile> pdfFiles = new List<PdfFile>();
 
-        static List<TextFile> textFiles = new List<TextFile>();
+        private static List<TextFile> textFiles = new List<TextFile>();
 
-        static System.Collections.Concurrent.ConcurrentBag<TaggedFile> taggedFiles = new System.Collections.Concurrent.ConcurrentBag<TaggedFile>();
-
-        public static List<TaggedFile> TaggedFiles {
-            get {
-                return FileManager.taggedFiles.ToList();
-            }
-            set {
-                foreach (var f in value) {
-                    FileManager.taggedFiles.Add(f);
-                }
-            }
-        }
+        private static List<TaggedFile> taggedFiles = new List<TaggedFile>();
 
         #endregion
     }
-    #region Exception Types
 
-
-    [Serializable]
-    class UnsupportedFileTypeAddedException : FileSystemException
-    {
-        public UnsupportedFileTypeAddedException(string unsupportedFormat)
-            : this(unsupportedFormat, null) {
-        }
-        public UnsupportedFileTypeAddedException(string unsupportedFormat, Exception inner)
-            : base(
-            String.Format(
-            "Files of type \"{0}\" are not supported. Supported types are {1}, {2}, {3}, and {4}",
-            unsupportedFormat,
-            from k in FileManager.WrapperMap.Keys.Take(4)
-            select k), inner) {
-
-        }
-
-        public UnsupportedFileTypeAddedException(System.Runtime.Serialization.SerializationInfo info, System.Runtime.Serialization.StreamingContext context)
-            : base(info, context) {
-        }
-    }
     #region Helper Types
     public class WrapperDict : Dictionary<string, Func<string, InputFile>>
     {
@@ -630,8 +654,43 @@ namespace LASI.FileSystem
     }
 
     #endregion
+
+    #region Exception Types
     [Serializable]
-    class FileManagerException : FileSystemException
+    public class FileManagerNotInitializedException : FileManagerException
+    {
+        public FileManagerNotInitializedException()
+            : base("File Manager has not been initialized. No directory context in which to operate.") {
+        }
+        public FileManagerNotInitializedException(System.Runtime.Serialization.SerializationInfo info, System.Runtime.Serialization.StreamingContext context)
+            : base(info, context) {
+        }
+    }
+
+    [Serializable]
+    public class UnsupportedFileTypeAddedException : FileManagerException
+    {
+        public UnsupportedFileTypeAddedException(string unsupportedFormat)
+            : this(unsupportedFormat, null) {
+        }
+        public UnsupportedFileTypeAddedException(string unsupportedFormat, Exception inner)
+            : base(
+            String.Format(
+            "Files of type \"{0}\" are not supported. Supported types are {1}, {2}, {3}, and {4}",
+            unsupportedFormat,
+            from k in FileManager.WrapperMap.Keys.Take(4)
+            select k), inner) {
+
+        }
+
+        public UnsupportedFileTypeAddedException(System.Runtime.Serialization.SerializationInfo info, System.Runtime.Serialization.StreamingContext context)
+            : base(info, context) {
+        }
+    }
+
+
+    [Serializable]
+    public class FileManagerException : FileSystemException
     {
 
         protected FileManagerException(string message)
@@ -645,7 +704,7 @@ namespace LASI.FileSystem
         }
 
 
-        public FileManagerException(System.Runtime.Serialization.SerializationInfo info, System.Runtime.Serialization.StreamingContext context)
+        protected FileManagerException(System.Runtime.Serialization.SerializationInfo info, System.Runtime.Serialization.StreamingContext context)
             : base(info, context) {
             CollectDirInfo();
         }
@@ -668,7 +727,7 @@ namespace LASI.FileSystem
 
     }
     [Serializable]
-    abstract class FileSystemException : Exception
+    public abstract class FileSystemException : Exception
     {
 
         protected FileSystemException(string message)
